chore: specify types needed directly.
I'm not sure getting them through v10 of Replicache
was doing the right thing. When I install replicache-react and
replicache v12 together in an app, I only see one version of
replicache, v12. And replicache-react doesn't re-include the
required types in its own .d.ts file. So unless tsc is being
very fancy (which, I dunno, maybe it is) and getting the old
version of replicache behind the scenes somehow, then I think
this is using the wrong types.

Anyway, this is more direct way to do what was meant.