Remove duplicate render in test.
OK so I still think the implementation of `useSubscribe` is fine.
What was happening with the test was that `setState` *does* skip
re-renders if the value is identical via Object.is().

When I changed the semantics of useSubscribe() I didn't change
some of the test code to match. Previously the test code was
manually coallescing undefined to null. Because the snapshot
state inside useSubscribe now starts out as undefined rather
than `def` this meant that when the query returned, it caused a
re-render.